requirements.txt
alembic==1.1.0
blinker==1.4
Click==7.0
Faker==2.0.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
flask-redis==0.4.0
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.0.3
PyMySQL==0.9.3
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
redis==3.3.8
six==1.12.0
SQLAlchemy==1.3.8
text-unidecode==1.2
Werkzeug==0.15.6
WTForms==2.2.1
